Update patch set 2
Patch Set 2:

(1 comment)

> If this package is replacing openstack-heat-templates, should it have a Provides statement for backwards compatibility?, i'm not sure how it gets installed when needed.

The subpackages are the important thing, and their names have not changed. The new parent package contains no files and requires all of the subpackages. The old parent package didn't contain any files either, but nor did it require all of the subpackages. For that reason we decided that adding Provides would not be appropriate in this case.

> Is this new package required in ocata cycle?

Yes.

Patch-set: 2
Label: Code-Review=0