Update patch set 1
Patch Set 1:

> unless this is in ubuntu packaging for heat-agents, i doubt anyone is using it.

Because... only Ubuntu is used as a guest on OpenStack? I don't understand what you're trying to say here.
(FWIW I don't think heat-agents is packaged at all for Debian, so people have to use the d-i-b elements to use them on Ubuntu.)

> FYI Zane, we don't plan to maintain Paunch for long term

That's a problem, as we all discussed in a slightly different context here: http://eavesdrop.openstack.org/irclogs/%23heat/%23heat.2019-08-23.log.html#t2019-08-23T01:52:20

Unless somebody can explain to me why I am crazy, I am still of the same opinion.

Maybe it would make sense for the Heat team (such as it is) to take ownership of the paunch project if TripleO intends to drop it.

Patch-set: 1